requirements.txt
dj-database-url==0.5.0
Django==4.2.3
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
wheel==0.37.1
whitenoise==6.5.0
djangorestframework==3.15.1
django-cors-headers==4.3.1